+++ Source/EventFlow.RabbitMQ.Tests/Integration/RabbitMqTests.cs
@@ -54,7 +54,7 @@ namespace EventFlow.RabbitMQ.Tests.Integration
             var url = Environment.GetEnvironmentVariable("RABBITMQ_URL");
             if (string.IsNullOrEmpty(url))
             {
-                Assert.Inconclusive("The environment variabel named 'RABBITMQ_URL' isn't set. Set it to e.g. 'amqp://localhost'");
+                Assert.Inconclusive("The environment variable named 'RABBITMQ_URL' isn't set. Set it to e.g. 'amqp://localhost'");
             }
 
             _uri = new Uri(url);