+++ RELEASE_NOTES.md
@@ -1,6 +1,14 @@
 ### New in 0.55 (not released yet)
 
-* _Nothing yet_
+* New: Trigger sagas without the need of any domain events by using
+  `ISagaStore.UpdateAsync(...)`
+* New: .NET standard 2.0 (still supports 1.6) support added to these
+  NuGet packages
+  - EventFlow
+  - EventFlow.Autofac
+  - EventFlow.Elasticsearch
+  - EventFlow.Hangfire
+  - EventFlow.Sql
 
 ### New in 0.54.3261 (released 2018-02-25)